Time: 0 ms (100.00%), Space: 6.3 MB (58.39%) - LeetHub

+++ 0191-number-of-1-bits/0191-number-of-1-bits.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int hammingWeight(uint32_t n) {
-        int cnt = 0;
+        short cnt = 0;
         while(n > 0){
             if(n%2) cnt++;
             n/=2;
